Move duplicateAndDragCallback into scratchBlocksUtils
