+++ core/contextmenu.js
@@ -31,6 +31,7 @@
 goog.provide('Blockly.ContextMenu');
 
 goog.require('Blockly.Events.BlockCreate');
+goog.require('Blockly.scratchBlocksUtils');
 goog.require('Blockly.utils');
 goog.require('Blockly.utils.uiMenu');
 
@@ -264,7 +265,7 @@ Blockly.ContextMenu.blockDuplicateOption = function(block) {
   var duplicateOption = {
     text: Blockly.Msg.DUPLICATE,
     enabled: true,
-    callback: block.duplicateAndDragCallback_()
+    callback: Blockly.scratchBlocksUtils.duplicateAndDragCallback(block)
   };
   return duplicateOption;
 };
